requirements.txt
kopf~=0.21
kubernetes~=10.0
pyyaml~=5.1
google-auth~=1.6
jsonpath-ng~=1.4
